+++ AmazonAdMarketplace/CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
-## x.x.x.x
+## 4.3.1.1
+* Add support for updating cached ad loaders with new ones passed in via local parameters.
+
+## 4.3.1.0
+* Certified with APS SDK 4.3.1.
+
+## 4.3.0.0
 * Update open source versions to allow compilation with AppLovin SDK v11.0.0+.
+* Certified with APS SDK 4.3.0.
 
 ## 4.2.1.0
 * Initial commit.